refactor(daemon): extract QueueCommand from Commands::Daemon (#254)
Closes #254 — the final PR #244 /ce-code-review follow-up. Moves the 9
queue_* methods (which touch only @queue_args/@json/@hive_home) into
Hive::Commands::Daemon::QueueCommand, mirroring the ServiceInstaller
extraction. Commands::Daemon#queue_command lazily requires and delegates.

Stacked on #294 (fix/daemon-queue-244-followups-a): the moved queue_prune
already carries #265's remove_if_unclaimed. The extracted call also folds
forward #262's Hive::InternalError wrapping (exit 70 on internal failure,
from #297) so a rebase onto a main with #297 can't drop it. Must merge
after #294, #295, #297.

100% line coverage, rubocop clean.
